backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
google-generativeai
python-dotenv
requests
